+++ .gitignore
@@ -1,1 +1,2 @@
 output/*.json
+__pycache__
